Refactor and create useSiteActions
